fix bu

+++ app/Http/Controllers/Api/V1/RingController.php
@@ -188,25 +188,27 @@ class RingController extends Controller
             }
             $r->fill($data)->save();
             $memImage = $request->file('members');
-            foreach ($request->input('members') as $k => $member) {
-                $m = ! empty($member['id']) ? RingMember::query()->with(['image'])->find($member['id']) : new RingMember;
-                if (! empty($member['birthdate'])) {
-                    $member['birthdate'] =  dateConverter($member['birthdate'] ,'m');
-                }
-                $m->fill($member);
-                $m->ring()->associate($r);
-                $m->save();
-                if (! empty($memImage[$k]['image']) && $memImage[$k]['image'] instanceof UploadedFile) {
-                    $path =  '/rings/'.$now->year.'/'.$now->month.'/'.$now->day;
-                    $m->image()->create([
-                        'path' => $memImage[$k]['image']->store($path,$disk),
-                        'mime_type' => $memImage[$k]['image']->getMimeType(),
-                        'size' => $memImage[$k]['image']->getSize(),
-                        'disk' => $disk,
-                        'subject' => $m::FILE_IMAGE_SUBJECT
-                    ]);
-                    if ($m->image) {
-                        $m->image->delete();
+            if ($request->filled('members')) {
+                foreach ($request->input('members') ?? []  as $k => $member) {
+                    $m = ! empty($member['id']) ? RingMember::query()->with(['image'])->find($member['id']) : new RingMember;
+                    if (! empty($member['birthdate'])) {
+                        $member['birthdate'] =  dateConverter($member['birthdate'] ,'m');
+                    }
+                    $m->fill($member);
+                    $m->ring()->associate($r);
+                    $m->save();
+                    if (! empty($memImage[$k]['image']) && $memImage[$k]['image'] instanceof UploadedFile) {
+                        $path =  '/rings/'.$now->year.'/'.$now->month.'/'.$now->day;
+                        $m->image()->create([
+                            'path' => $memImage[$k]['image']->store($path,$disk),
+                            'mime_type' => $memImage[$k]['image']->getMimeType(),
+                            'size' => $memImage[$k]['image']->getSize(),
+                            'disk' => $disk,
+                            'subject' => $m::FILE_IMAGE_SUBJECT
+                        ]);
+                        if ($m->image) {
+                            $m->image->delete();
+                        }
                     }
                 }
             }
@@ -238,7 +240,7 @@ class RingController extends Controller
         $r = Ring::query()->with(['members'])->where('owner_id' , \auth()->id())->find($ring);
 
         try {
-            foreach ($r->members as $member) {
+            foreach ($r->members ?? [] as $member) {
                 if ($member->image) {
                     $member->image->delete();
                 }
